refactor: 예외처리 경로 변경 리팩토링
