+++ src/main/java/hello/backend/image/service/FileStorageService.java
@@ -1,6 +1,6 @@
 package hello.backend.image.service;
 
-import hello.backend.exception.InvalidFileException;
+import hello.backend.error.exception.user.InvalidFileException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
